feat(user): Añadir endpoint de creación de usuarios desde API externa.

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/nequi/r2dbc/MyReactiveRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 // TODO: This file is just an example, you should delete or modify it
-public interface MyReactiveRepository extends ReactiveCrudRepository<UserEntity, Integer>,
+public interface MyReactiveRepository extends ReactiveCrudRepository<UserEntity, Long>,
         ReactiveQueryByExampleExecutor<UserEntity> {
 
 }
